# ecommerce-controlled/order-service/src/main/java/com/ecommerce/order/dto/OrderItemResponse.java
package com.ecommerce.order.dto;

import java.math.BigDecimal;
import java.util.UUID;

public record OrderItemResponse(
    UUID id,
    String productId,
    String productName,
    Integer quantity,
    BigDecimal unitPrice,
    BigDecimal totalPrice
) {
}